--- monitoring-data-with-react/src/Components/chartandgraph/Table.jsx
@@ -1,165 +1,0 @@
-import React, { useState, useEffect } from "react";
-import { motion } from "framer-motion";
-import { Search } from "lucide-react";
-import DatePicker from "react-datepicker";
-import "react-datepicker/dist/react-datepicker.css";
-import { parseDataToTable } from "../../utils/parseDataToTable";
-const OrdersTable = ({ title }) => {
-  const [searchTerm, setSearchTerm] = useState("");
-  const [dataResults, setDataResults] = useState([]);
-  const [filteredOrders, setFilteredOrders] = useState([]);
-  const [dateRange, setDateRange] = useState([null, null]);
-  const [startDate, endDate] = dateRange;
-  const [time, setTime] = useState("10m");
-
-  useEffect(() => {
-    const fetchData = async () => {
-      const data = await parseDataToTable(time, startDate, endDate);
-      setDataResults(data);
-      filterOrders(searchTerm, data);
-    };
-    fetchData();
-  }, [time, startDate, endDate]);
-
-  const handleSearch = (e) => {
-    const term = e.target.value.toLowerCase();
-    console.log(term);
-    setSearchTerm(term);
-    filterOrders(term, dataResults);
-  };
-
-  const handleDateChange = (dates) => {
-    setDateRange(dates);
-  };
-
-  const handleChangeTime = (e) => {
-    e.preventDefault();
-    setTime(e.target.value);
-  };
-
-  // const filterOrders = (term, data) => {
-  //   const filtered = data.filter((order) => {
-  //     const matchesTerm = term
-  //       ? order.status.toLowerCase().includes(term) ||
-  //         order.date.toLowerCase().includes(term) ||
-  //         order.time.toLowerCase().includes(term)
-  //       : true;
-  //     return matchesTerm;
-  //   });
-  //   setFilteredOrders(filtered);
-  // };
-  const filterOrders = (term, data) => {
-    const filtered = data.filter((order) => {
-      const matchesTerm = term
-        ? order.status.toLowerCase().includes(term) ||
-          order.date.toLowerCase().includes(term) ||
-          order.time.toLowerCase().includes(term)
-        : true;
-      return matchesTerm;
-    });
-    const sortedFiltered = filtered.sort((a, b) => b.timestamp - a.timestamp);
-
-    setFilteredOrders(sortedFiltered);
-  };
-
-  return (
-    <motion.div
-      className="bg-gray-800 bg-opacity-50 backdrop-blur-md shadow-lg rounded-xl p-6 border border-gray-700"
-      initial={{ opacity: 0, y: 20 }}
-      animate={{ opacity: 1, y: 0 }}
-      transition={{ delay: 0.4 }}
-    >
-      <div className="flex justify-between items-center mb-6">
-        <h2 className="text-xl font-semibold text-gray-100">{title}</h2>
-        <div className="flex items-center space-x-4">
-          <DatePicker
-            selectsRange
-            startDate={startDate}
-            endDate={endDate}
-            onChange={handleDateChange}
-            placeholderText="Pilih rentang tanggal"
-            className="bg-gray-700 text-white placeholder-gray-400 rounded-lg pl-3 pr-4 py-2 focus:outline-none focus:ring-2 focus:ring-blue-500"
-          />
-          <select
-            className="bg-gray-700 text-white rounded-md mx-2 px-3 py-1 focus:outline-none focus:ring-2 focus:ring-blue-500"
-            onChange={handleChangeTime}
-            value={time}
-          >
-            <option value="60s">1 minute</option>
-            <option value="5m">5 minutes</option>
-            <option value="10m">10 minutes</option>
-            <option value="15m">15 minutes</option>
-            <option value="30m">30 minutes</option>
-            <option value="1h">1 hour</option>
-          </select>
-          <div className="relative">
-            <input
-              type="text"
-              placeholder="Cari pesanan..."
-              className="bg-gray-700 text-white placeholder-gray-400 rounded-lg pl-10 pr-4 py-2 focus:outline-none focus:ring-2 focus:ring-blue-500"
-              value={searchTerm}
-              onChange={handleSearch}
-            />
-            <Search
-              className="absolute left-3 top-2.5 text-gray-400"
-              size={18}
-            />
-          </div>
-        </div>
-      </div>
-
-      <div className="overflow-x-auto">
-        <table className="min-w-full divide-y divide-gray-700">
-          <thead>
-            <tr>
-              <th className="px-6 py-3 text-left text-xs font-medium text-gray-400 uppercase tracking-wider">
-                Status
-              </th>
-              <th className="px-6 py-3 text-left text-xs font-medium text-gray-400 uppercase tracking-wider">
-                Date
-              </th>
-              <th className="px-6 py-3 text-left text-xs font-medium text-gray-400 uppercase tracking-wider">
-                Time
-              </th>
-            </tr>
-          </thead>
-
-          <tbody className="divide-y divide-gray-700">
-            {filteredOrders.map((order, index) => (
-              <motion.tr
-                key={index}
-                initial={{ opacity: 0 }}
-                animate={{ opacity: 1 }}
-                transition={{ duration: 0.3 }}
-              >
-                <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-300">
-                  <span
-                    className={`px-2 inline-flex text-xs leading-5 font-semibold rounded-full ${
-                      order.status === "Delivered"
-                        ? "bg-green-100 text-green-800"
-                        : order.status === "Processing"
-                        ? "bg-yellow-100 text-yellow-800"
-                        : order.status === "Shipped"
-                        ? "bg-blue-100 text-blue-800"
-                        : "bg-red-100 text-red-800"
-                    }`}
-                  >
-                    {order.status}
-                  </span>
-                </td>
-                <td className="px-6 py-4 whitespace-nowrap text-sm font-medium text-gray-100">
-                  {order.date}
-                </td>
-                <td className="px-6 py-4 whitespace-nowrap text-sm font-medium text-gray-100">
-                  {order.time}
-                </td>
-              </motion.tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
-    </motion.div>
-  );
-};
-
-export default OrdersTable;
